fix: harden crash report sanitizer and clean up assertions
- Cap exception message length at 4 KB in crash reports.
- Remove InvalidDataException from expected capture exceptions.
- Strengthen E2E smoke report file detection and extension assertion.
- Use CrashLogEvents constants instead of magic values in field policy tests.
- Share ManualTimeProvider from TestData and remove duplicate nested definitions.
- Fix breadcrumb timestamp-based assertions in text writer tests.
